+++ Sources/ZcashLightClientKit/Extensions/Data+internal.swift
@@ -19,6 +19,26 @@ extension String {
             }
         return id
     }
+    
+    func txIdToBytes() -> [UInt8] {
+        var bytes: [UInt8] = []
+        bytes.reserveCapacity(count / 2)
+        
+        var index = startIndex
+        
+        while index < endIndex {
+            let nextIndex = self.index(index, offsetBy: 2)
+            let byteString = self[index..<nextIndex]
+            if let byte = UInt8(byteString, radix: 16) {
+                bytes.append(byte)
+            } else {
+                return [] // or throw error
+            }
+            index = nextIndex
+        }
+        
+        return bytes
+    }
 }
 
 extension Collection {